+++ src/renderer/src/App.tsx
@@ -17,6 +17,7 @@ import { NewPromptModal } from './components/modals/NewPromptModal';
 import { AddOutputModal } from './components/modals/AddOutputModal';
 import { NewModelModal } from './components/modals/NewModelModal';
 import { RunBenchmarkModal } from './components/modals/RunBenchmarkModal';
+import { CommandPalette } from './components/common/CommandPalette';
 import { CheckCircle2, AlertCircle, Info, X } from 'lucide-react';
 
 export const MainApp: React.FC = () => {
@@ -71,6 +72,7 @@ export const MainApp: React.FC = () => {
       <StatusBar />
 
       {/* Global Modals */}
+      <CommandPalette />
       <NewPromptModal />
       <AddOutputModal />
       <NewModelModal />